+++ CHANGELOG.md
@@ -87,6 +87,8 @@ Semantic versioning in our case means:
 - Allows class / instance attribute shadowing
   in `@dataclass`es for `WPS601`, #1926
 - Allows any number of instance attributes on `@dataclass`es in `WPS230`, #2448
+- Allows any number of function parameters
+  in `@overload` definitions for `WPS211`, #1957
 - Adds new rule to forbid `lambda` assigns to special attributes, #1733
 - Adds new rule to check problematic function params, #1343
 - Adds new rule to detect duplicate conditions in `if`s and `elif`s, #2241